+++ front-end/package.json
@@ -14,6 +14,7 @@
     "postcss": "^8.4.35",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-image-crop": "^11.0.5",
     "react-markdown": "^9.0.1",
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
